requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.13.4
docling==2.34.0
fastapi==0.111.1
langchain-core==0.3.61
langchain-text-splitters==0.3.8
langdetect==1.0.9
markdownify==0.13.1
numpy==2.2.6
pydantic==2.11.5
pypdf==4.3.1
python-docx==1.1.2
python-dotenv==1.1.0
qdrant-client==1.14.2
requests==2.32.3
scikit-learn==1.5.1
sentence-transformers==4.1.0
setuptools==80.8.0
spacy==3.8.6
starlette==0.37.2
tiktoken==0.6.0
uvicorn==0.29.0
wasabi==1.1.2
weaviate-client==4.9.6
